Struct xxhash2::Hash64

pub struct Hash64 {
    // some fields omitted
}

Canonical representation of a 64-bit hash.

This structure provides a conversion from a 64-bit hash value to a list of bytes in a canonical format. Note that the bytes returned are not necessarily hex.

Methods

impl Hash64

fn bytes(&self) -> &[u8; 8]

Returns the underlying hash as a list of bytes in a canonical representation.

fn value(&self) -> u64

Returns this canonical hash value as a 64-bit integer.

Converts from the underlying list of bytes to an integer.