Struct xxhash2::State32

pub struct State32 {
    // some fields omitted
}

Representation of the intermediate state of a 32-bit xxHash instance.

This structure can be used to generate a 32-bit hash of a block of bytes.

Methods

impl State32

fn new() -> State32

Creates a new blank instance of the 32-bit xxHash state.

This state can then be used to hash a list of bytes and acquire the result via the finish method.

Examples

let state = xxhash2::State32::new();

fn update(&mut self, data: &[u8])

Input a block of bytes into this hasher.

This function will update the internal state of this hasher with a new list of bytes to feed in. To get the hash value of the block (and all previous bytes), call the finish function.

Examples

let mut state = xxhash2::State32::new();
state.update(&[1, 2, 3, 4]);

fn reset(&mut self, seed: u32)

Reset the internal state of this hasher with a given seed.

This is useful to reuse a hasher or to persist the state across multiple runs of a hasher.

Examples

let mut state = xxhash2::State32::new();
state.update(&[1, 2, 3, 4]);

fn finish(&self) -> u32

Computes the final hash result of all bytes that have been input to this hasher.

Returns the 32-bit checksum of the result.

Examples

let mut state = xxhash2::State32::new();
state.update(&[1, 2, 3, 4]);
assert_eq!(state.finish(), 4271296924);